+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Refactoring
 - Persistent OIDC authorization codes (#25, PLUM Sprint 220408)
 - Persistent TOTP secrets (#27, PLUM Sprint 220408)
+- Revise default configs and examples (#28, PLUM Sprint 220422)
 
 ---