+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Release candidate
 
+### Fix
+- Use datetime.datetime.utcnow (#29, PLUM Sprint 220422)
+
 ### Refactoring
 - Persistent OIDC authorization codes (#25, PLUM Sprint 220408)
 - Persistent TOTP secrets (#27, PLUM Sprint 220408)