+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Persistent TOTP secrets (#27, PLUM Sprint 220408)
 - Revise default configs and examples (#28, PLUM Sprint 220422)
 - Persistent login sessions (#26, PLUM Sprint 220422)
+- Session adapter restructured (#32, PLUM Sprint 220506)
 
 ---