+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Refactoring
 - Persistent OIDC authorization codes (#25, PLUM Sprint 220408)
 - Persistent TOTP secrets (#27, PLUM Sprint 220408)
+- Dockerfile is now building in two stages
 - Revise default configs and examples (#28, PLUM Sprint 220422)
 - Persistent login sessions (#26, PLUM Sprint 220422)
 - Session adapter restructured (#32, PLUM Sprint 220506)