+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Release candidate
 
+
+## v22.21
+
 ### Fix
 - Use datetime.datetime.utcnow (#29, PLUM Sprint 220422)
 - After-provisioning cleanup (#33, PLUM Sprint 220506)